Remove BaseActivity as only the MainActivity should be able to load the settings panel
