--- app/src/main/java/io/github/imsmobile/fahrplan/BaseActivity.java
@@ -1,45 +1,0 @@
-package io.github.imsmobile.fahrplan;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.annotation.LayoutRes;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-
-
-
-public abstract class BaseActivity  extends AppCompatActivity {
-
-    public void onCreate(Bundle savedInstanceState, @LayoutRes int layoutResID) {
-        super.onCreate(savedInstanceState);
-        setContentView(layoutResID);
-        setSupportActionBar((Toolbar) findViewById(R.id.toolbar));
-    }
-
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.toolbar, menu);
-        return super.onCreateOptionsMenu(menu);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_settings:
-                startSettingActivity();
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
-    }
-
-    private void startSettingActivity() {
-        Intent intent = new Intent(this, SettingActivity.class);
-        startActivity(intent);
-    }
-}
