+++ app/src/main/java/io/github/imsmobile/fahrplan/MainActivity.java
@@ -2,22 +2,26 @@ package io.github.imsmobile.fahrplan;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 
-public class MainActivity extends BaseActivity{
+public class MainActivity extends AppCompatActivity {
     public final static String FROM_MESSAGE = "io.github.imsmoble.fahrplan.from";
     public final static String TO_MESSAGE = "io.github.imsmoble.fahrplan.to";
 
     @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState, R.layout.activity_main);
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        setSupportActionBar((Toolbar) findViewById(R.id.toolbar));
         registerSearchButton();
     }
-
-
 
     private void registerSearchButton() {
         Button buttonOne = (Button) findViewById(R.id.btn_search);
@@ -37,6 +41,29 @@ public class MainActivity extends BaseActivity{
         String to = toText.getText().toString();
         intent.putExtra(FROM_MESSAGE, from);
         intent.putExtra(TO_MESSAGE, to);
+        startActivity(intent);
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.toolbar, menu);
+        return super.onCreateOptionsMenu(menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case R.id.action_settings:
+                startSettingActivity();
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+    }
+
+    private void startSettingActivity() {
+        Intent intent = new Intent(this, SettingActivity.class);
         startActivity(intent);
     }
 }
